requirements.txt
beautifulsoup4==4.14.3
pymysql==1.1.2
python-dotenv==1.2.1
requests==2.32.5
